Feat(debug): Adding a debug directory
I choose to add a debug directory, in the master branch and not in a debug branch during the development process.
Because if I need to intervene in the database in an emergency during a competition, I want to be able to find the problem as quickly as possible with this debugging folder.
It is a personal choice, I will separate this folder at the end of the production part of the application.